requirements.txt
mcp
flask==2.2.5
flask-cors
werkzeug==2.2.3
requests
Pillow
langchain-groq
python-dotenv
groq
PyPDF2
python-docx
pydantic
aiohttp
fastmcp
langdetect
redis
SQLAlchemy==1.4.53
flask-login
flask-migrate
flask-wtf
flask-sqlalchemy==2.5.1
numpy
packaging
google-generativeai
jsonschema
httpx
fastapi
uvicorn
psycopg2-binary
llm-guard
langchain
#Install separately , one by one. Also use uv ! its super fast !
rasa
rasa"[spacy]"mcp
flask==2.2.5
flask-cors
werkzeug==2.2.3
requests
Pillow
langchain-groq
python-dotenv
groq
PyPDF2
python-docx
pydantic
aiohttp
fastmcp
langdetect
redis
SQLAlchemy==1.4.53
flask-login
flask-migrate
flask-wtf
flask-sqlalchemy==2.5.1
numpy
packaging
google-generativeai
jsonschema
httpx
fastapi
uvicorn
psycopg2-binary
llm-guard
langchain
#Install separately , one by one. Also use uv ! its super fast !
rasa
rasa"[spacy]"